+++ node/src/types/kernel/api/block/getChildBlocks/payload.d.ts
@@ -15,10 +15,12 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-
-/* directories */
-export * as getChildBlocks from "./getChildBlocks";
-
-/* files */
-export * as getBlockBreadcrumb from "./getBlockBreadcrumb";
-export * as getDocInfo from "./getDocInfo";
+/**
+ * Get child blocks
+ */
+export interface IPayload {
+    /**
+     * block ID
+     */
+    id: string;
+}