+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - 添加 API `/api/filetree/getHPathByID` | Add API `/api/filetree/getHPathByID`
 - 添加 API `/api/filetree/moveDocs` | Add API `/api/filetree/moveDocs`
 - 添加 API `/api/asset/upload` | Add API `/api/asset/upload`
+- 添加 API `/api/block/getChildBlocks` | Add API `/api/block/getChildBlocks`
 
 ## v0.0.5 / 2023-06-30